+++ src/services/balanceMonitor/BalanceMonitorKeyResolver.ts
@@ -1,0 +1,7 @@
+import {ChainsIds} from "../../types/ChainsIds";
+
+export class BalanceMonitorKeyResolver {
+  static apply(chainId: ChainsIds, wallet: string): string {
+    return `BALANCE_MONITOR_${chainId.toUpperCase()}_${wallet.toUpperCase()}`;
+  }
+}